# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Creators Dashboard classifier API
  description: |
    A Flask + MongoDB API for the classifier service.
    
    ## Authentication
    All endpoints (except `/metrics`) require JWT authentication via the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    
    For local development, obtain a JWT from the Developer Edition welcome / IdP flow, or use a
    test token from `test/e2e/e2e_auth.py` that matches your `JWT_SECRET` and expected claims.
  version: 0.1.0
  contact:
    name: Agile Crafts People
    email: info@agilecrafts.com
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:9100
    description: Local development server
...
paths:
...
  /api/User:
    get:
      tags:
        - User
      summary: Get all User documents
      description: Retrieve infinite scroll batch of sorted, filtered User documents (read-only)
      operationId: getUsers
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-User
        - name: after_id
          in: query
          required: false
          description: Cursor for infinite scroll (ID of last item from previous batch, omit for first request)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
        - name: limit
          in: query
          required: false
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (name, description)
          schema:
            type: string
            default: name
            example: name
        - name: order
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            example: asc
      responses:
        '200':
          description: Successfully retrieved Users
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - limit
                  - has_more
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  limit:
                    type: integer
                    example: 10
                  has_more:
                    type: boolean
                    example: true
                  next_cursor:
                    type: string
                    nullable: true
                    pattern: '^[0-9a-fA-F]{24}$'
                    example: 507f1f77bcf86cd799439011
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/User/{UserId}:
    get:
      tags:
        - User
      summary: Get User by UserId
      description: Retrieve a specific User document by its UserId (read-only)
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - name: UserId
          in: path
          required: true
          description: The User ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by your identity provider (Developer Edition welcome flow for local dev).
        Include the token in the Authorization header: `Bearer <token>`
...
  responses:
    BadRequest:
      description: Bad Request - Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Unauthorized
...
    User:
      type: object
      description: User domain document (read-only)
      additionalProperties: false
      required:
        - _id
        - name
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: User name
          pattern: '^[^\s]{1,40}$'
          example: my-User
        description:
          type: string
          description: User description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample User document
        status:
          type: string
          description: User status
          example: active
...